feat(events): abort stalled preflight gates
